+++ repro/models/dugan2020.py
@@ -1,0 +1,2 @@
+from ._models.dugan2020.src.model import RoFTRecipeGenerator
+from ._models.dugan2020.src.setup import Dugan2020SetupSubcommand