+++ repro/models/model.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, List
+from typing import Any, Dict, List, Union
 
 from repro.common import Registrable
 from repro.data.types import DocumentType, SummaryType
@@ -29,6 +29,16 @@ class QuestionAnsweringModel(Model):
         return self.predict_batch([{"context": context, "question": question}])[0]
 
     def predict_batch(self, inputs: List[Dict[str, str]], *args, **kwargs) -> List[str]:
+        raise NotImplementedError
+
+
+class RecipeGenerationModel(Model):
+    def predict(self, name: str, ingredients: List[str], *args, **kwargs) -> str:
+        return self.predict_batch([{"name": name, "ingredients": ingredients}])[0]
+
+    def predict_batch(
+        self, inputs: List[Dict[str, Union[str, List[str]]]], *args, **kwargs
+    ) -> List[str]:
         raise NotImplementedError